requirements.txt
streamlit==1.24.0
pandas==1.3.5
numpy==1.21.6
matplotlib==3.5.3
seaborn==0.12.0
scikit-learn==1.0.2
xgboost==1.6.2
shap==0.41.0
plotly==5.10.0
pillow==9.2.0
joblib==1.1.0
scipy==1.7.3
threadpoolctl==3.1.0
llvmlite==0.39.0
numba==0.56.4
protobuf==3.20.3
requests==2.27.1
beautifulsoup4==4.10.0
selenium==4.1.0
webdriver-manager==3.5.2 